requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai>=0.8.3
pandas
numpy
pydantic
jinja2
python-multipart